test(policy): move fuzz tests from core to policy and add unit tests (#2196)
